requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.32,<1
python-multipart>=0.0.9
jinja2>=3.1,<4
pydantic-settings>=2.0,<3
torch
torchvision
transformers==4.45.2
kornia
pillow; platform_system == "Windows"
pillow-simd; platform_system != "Windows"
huggingface_hub
timm
python-dotenv
pikepdf>=10.0,<11
pypdf>=5.0,<7
pypdfium2>=4.30,<5
img2pdf>=0.5,<1
